Make CMakeLists.txt cleaner
One of the notable changes is that new CMakeLists do not try to set any
extra compilation flags anymore. Realistically, this project won't be
used outside kege-by-danshaders in a foreseeable future, and KEGE will
soon have a centralized location for compilation options.

Additionally, this removes clang-format target, which should not have
existed in the first place.